UrbanScienceScraper/TimerTrigger_UrbanScience/timer_function.py
import datetime
import logging
import azure.functions as func

def main(mytimer: func.TimerRequest) -> None:
    utc_timestamp = datetime.datetime.utcnow().replace(tzinfo=datetime.timezone.utc).isoformat()
    logging.info(f"Timer trigger function ran at {utc_timestamp}")